// src/modules/property/types/properties.ts
export interface PropertyItem {
  id: string;
  propertyCode: string;
  name: string;
  location: string;
  propertyType: string;
  landArea: string;
  builtUpArea: string;
  totalFloors: string;
  totalUnits: number;
  availableUnits: number;
  occupancyRate: string;
  rajukApprovalNo: string;
  marketValuation: string;
  completionYear: string;
  status: string;
}